I've given this some thought, and I'm “starting from scratch” so to speak with the new theme system -- i.e. I'm not going to try to “import” what changes the user may have done in the old theme editor.
The file format of a settings group is however the same as that of a theme. So if you've created your own settings groups (color sets), you can just copy them from Settings to Themes (and change extension to plist), and the color set will then show up as a theme (in 1.1b15).

git-svn-id: http://svn.textmate.org/trunk/Themes@1358 dfb7d73b-c2ec-0310-8fea-fb051d288c6d